Move merge operator tests from db_test.go to merge_test.go file
